Moved network response files to models
