+++ app/src/main/java/com/projects/moviemanager/domain/models/content/ContentCast.kt
@@ -1,6 +1,6 @@
 package com.projects.moviemanager.domain.models.content
 
-import com.projects.moviemanager.network.response.content.common.ContentCastResponse
+import com.projects.moviemanager.network.models.content.common.ContentCastResponse
 
 data class ContentCast(
     val id: Int,
