+++ app/src/main/java/com/projects/moviemanager/domain/models/content/MediaContentDetails.kt
@@ -1,10 +1,10 @@
 package com.projects.moviemanager.domain.models.content
 
 import com.projects.moviemanager.common.domain.MediaType
-import com.projects.moviemanager.network.response.content.common.ContentGenre
-import com.projects.moviemanager.network.response.content.common.ProductionCountry
-import com.projects.moviemanager.network.response.content.common.BaseMediaContentResponse
-import com.projects.moviemanager.network.response.person.PersonDetailsResponse
+import com.projects.moviemanager.network.models.content.common.ContentGenre
+import com.projects.moviemanager.network.models.content.common.ProductionCountry
+import com.projects.moviemanager.network.models.content.common.BaseMediaContentResponse
+import com.projects.moviemanager.network.models.person.PersonDetailsResponse
 sealed class DetailedMediaInfo : BaseMediaContent
 
 data class MovieDetailsInfo(
