+++ app/src/main/java/com/projects/moviemanager/features/details/domain/DetailsInteractor.kt
@@ -19,9 +19,9 @@ import com.projects.moviemanager.features.watchlist.model.DefaultLists
 import com.projects.moviemanager.network.repository.movie.MovieRepository
 import com.projects.moviemanager.network.repository.person.PersonRepository
 import com.projects.moviemanager.network.repository.show.ShowRepository
-import com.projects.moviemanager.network.response.content.movie.MovieApiResponse
-import com.projects.moviemanager.network.response.content.show.ShowApiResponse
-import com.projects.moviemanager.network.response.person.PersonDetailsResponse
+import com.projects.moviemanager.network.models.content.movie.MovieApiResponse
+import com.projects.moviemanager.network.models.content.show.ShowApiResponse
+import com.projects.moviemanager.network.models.person.PersonDetailsResponse
 import com.projects.moviemanager.network.util.Left
 import com.projects.moviemanager.network.util.Right
 import javax.inject.Inject
