+++ app/src/main/java/com/projects/moviemanager/network/models/content/common/ContentCreditsResponse.kt
@@ -1,4 +1,4 @@
-package com.projects.moviemanager.network.response.content.common
+package com.projects.moviemanager.network.models.content.common
 
 data class ContentCreditsResponse(
     val id: Int = 0,
