+++ app/src/main/java/com/projects/moviemanager/network/models/content/common/ProductionCountry.kt
@@ -1,4 +1,4 @@
-package com.projects.moviemanager.network.response.content.common
+package com.projects.moviemanager.network.models.content.common
 
 data class ProductionCountry(
     val iso_3166_1: String?,
