+++ app/src/main/java/com/projects/moviemanager/network/models/content/common/VideosByIdResponse.kt
@@ -1,0 +1,8 @@
+package com.projects.moviemanager.network.models.content.common
+
+import com.projects.moviemanager.network.models.content.common.VideoResponse
+
+data class VideosByIdResponse(
+    val id: Int?,
+    val results: List<VideoResponse>
+)
