+++ app/src/main/java/com/projects/moviemanager/network/repository/person/PersonRepository.kt
@@ -1,9 +1,9 @@
 package com.projects.moviemanager.network.repository.person
 
 import com.projects.moviemanager.network.models.ApiError
-import com.projects.moviemanager.network.response.person.PersonCreditsResponse
-import com.projects.moviemanager.network.response.person.PersonDetailsResponse
-import com.projects.moviemanager.network.response.person.PersonImagesResponse
+import com.projects.moviemanager.network.models.person.PersonCreditsResponse
+import com.projects.moviemanager.network.models.person.PersonDetailsResponse
+import com.projects.moviemanager.network.models.person.PersonImagesResponse
 import com.projects.moviemanager.network.util.Either
 import kotlinx.coroutines.flow.Flow
