--- app/src/main/java/com/projects/moviemanager/network/response/content/common/VideosByIdResponse.kt
@@ -1,8 +1,0 @@
-package com.projects.moviemanager.network.response.content.common
-
-import com.projects.moviemanager.network.response.content.common.VideoResponse
-
-data class VideosByIdResponse(
-    val id: Int?,
-    val results: List<VideoResponse>
-)
