+++ app/src/main/java/com/projects/moviemanager/network/services/person/PersonService.kt
@@ -1,8 +1,8 @@
 package com.projects.moviemanager.network.services.person
 
-import com.projects.moviemanager.network.response.person.PersonCreditsResponse
-import com.projects.moviemanager.network.response.person.PersonDetailsResponse
-import com.projects.moviemanager.network.response.person.PersonImagesResponse
+import com.projects.moviemanager.network.models.person.PersonCreditsResponse
+import com.projects.moviemanager.network.models.person.PersonDetailsResponse
+import com.projects.moviemanager.network.models.person.PersonImagesResponse
 import com.projects.moviemanager.util.Constants
 import com.projects.moviemanager.util.Constants.ENGLISH_LANGUAGE_CODE
 import retrofit2.Response
